kernel: finish implementing VMRegion for all relevant vmm ops
the virtmapper no longer uses the entry in the PML as the "gospel",
except when explicitly requested. this also means that we cannot
have a "freePhys" functionality when unmapping.

this functionality is now provided by VMRegion; vmm allocations
are tracked in the AddressSpace of the process, and each region is
freed (and the physical pages deallocated) when the process and by
extension the AddressSpace is destroyed.

this should make it more streamlined in most places...??