kernel: fix race condition on page fault causing Bad Stuff
threads would crash randomly with weird page faults, but only on
virtualbox; turns out the problem is in the page fault handler, and
the usage of an IST.

since we allow a faulting process to be preempted (in case it has
to fetch from disk or something), and the PF handler uses IST1, that
IST1 will be "held up" by that thread.

unfortunately when another thread page faults, it will re-use that
IST and trash the contents of the stack, so that when the original
thread wants to resume, it returns into a bogus context/stack frame.

massive oversight that should've been obvious in hindsight. this was
only happening on virtualbox because of our serial debug log in the
handler; on vbox, writing to serial is slow enough to open a window
for the thread to get pre-empted; on QEMU it happens to fast, and so
we never saw the bug.

big rip.