kernel: add collectOrReuseMemticket() function
as the name implies, if the memticket with the given id has been
collected before in the current process, this returns *AN* existing
collection (ptr/len), without incurring the cost of mapping virtual
memory. note that the specific collection returned is unspecified
if there is more than one in the current process.

if no existing collection in the current process exists, then this
function behaves identically to collectMemticket().