kernel: collectOrReuseMemticket -> findExistingMemticket
the former function was pretty dumb -- there wasn't a way for the
caller to know whether or not the memticket returned was reused or
not, and so it either (a) erroneously calls release() when reused,
or (b) leaks if it had to collect().

this new function will return an invalid memticket if a prior one
does not exist for reuse; this allows callers to determine if it
needs to call release().